+++ AlgorithmsVisualisation/GraphsAlgorithms/Metrics/Manhattan.cs
@@ -3,9 +3,9 @@ using GraphsAlgorithms.GraphModel;
 
 namespace GraphsAlgorithms.Metrics {
     public class Manhattan : IMetric {
-        public double VertexDiagonalDistance => 2.0;
+        public float VertexDiagonalDistance => 2.0f;
 
-        public double GetDistance(Vertex start, Vertex end) =>
+        public float GetDistance(Vertex start, Vertex end) =>
             Math.Abs(end.X - start.X) + Math.Abs(end.Y - start.Y);
     }
 }
